PR #700 improve tests for plugins, refactor moved code from WebMaster to FlowMaster
