+++ test/test_plugins.py
@@ -1,53 +1,48 @@
 import argparse
-from libmproxy import web, flow
+from libmproxy import script, flow
 import tutils
 from libmproxy.proxy.server import DummyServer
 from libmproxy.proxy import ProxyConfig
+import netlib.utils
+from netlib import odict
+from netlib.http.semantics import CONTENT_MISSING, HDR_FORM_URLENCODED, HDR_FORM_MULTIPART
 
 
-def test_web_plugins():
-    plugins = web.WebPlugins()
+def test_plugins():
+    plugins = flow.Plugins()
     assert plugins
 
     # two types: view and action
     assert len(dict(plugins).items()) == 2
 
 
-def test_web_plugins_views():
-    plugins = web.WebPlugins()
+def test_plugins_views():
+    plugins = flow.Plugins()
 
     # two types: view and action
-    assert len(dict(plugins).items()[1]) == 2
-    assert dict(plugins).items()[1][0] == 'view_plugins'
-    assert len(dict(plugins).items()[1][1]) == 0
+    assert 'view_plugins' in dict(plugins).keys()
+
+    view_plugins = plugins['view_plugins']
+    assert len(view_plugins) == 0
 
     plugins.register_view('noop',
                           title='Noop View Plugin',
                           transformer=lambda x: x)
 
-    assert len(dict(plugins).items()[1][1]) == 1
-    assert len(dict(plugins).items()[0][1]) == 0
+    assert len(view_plugins) == 1
+    assert view_plugins['noop']['title'] == 'Noop View Plugin'
 
 
-def test_web_plugins_actions():
-    r = tutils.treq()
-    f = tutils.tflow(resp=True)
-    s = flow.ServerPlaybackState(
-        None,
-        [],
-        False,
-        False,
-        None,
-        False,
-        None,
-        False)
-    fm = web.WebMaster(DummyServer(ProxyConfig()), web.Options())
+def test_plugins_actions():
+    s = flow.State()
+    fm = flow.FlowMaster(None, s)
+    sp = tutils.test_data.path("scripts/a.py")
+    f = tutils.tflow(req=netlib.tutils.treq(), resp=True)
 
     plugins = fm.plugins
 
-    assert len(dict(plugins).items()[0]) == 2
-    assert dict(plugins).items()[0][0] == 'action_plugins'
-    assert len(dict(plugins).items()[0][1]) == 0
+    assert 'action_plugins' in dict(plugins).keys()
+    action_plugins = dict(plugins)['action_plugins']
 
     plugins.register_action('noop',
                             title='noop',
@@ -69,26 +64,25 @@ def test_web_plugins_actions():
                             }
                             ])
 
-    assert len(dict(plugins).items()[0][1]) == 1
-    assert len(dict(plugins).items()[1][1]) == 0
+    assert len(action_plugins) == 1
 
     assert f.request.content == "content"
 
-    # now let's try one that should change it
-    plugins.register_action('test',
-                            title='test',
-                            actions=[{
-                                'title': 'test',
-                                'id': 'replace_with_test',
-                                'state': {
-                                    'every_flow': True,
-                                },
-                                'possible_hooks': ['request', 'response'],
-                            }
-                            ])
+
+def test_action_plugin_simple():
+    s = flow.State()
+    fm = flow.FlowMaster(None, s)
+    fm.load_script(tutils.test_data.path("scripts/a.py"))
+    f = tutils.tflow(req=netlib.tutils.treq(), resp=True)
+
+    plugins = fm.plugins
+    action_plugins = plugins['action_plugins']
+
+    assert len(action_plugins) == 1
+    assert f.request.content == 'content'
 
     fm.handle_clientconnect(f.client_conn)
     fm.handle_serverconnect(f.server_conn)
     fm.handle_request(f)
-    #fm.replay_request(f, run_scripthooks=True)
-    #assert f.request.content == 'test'
+
+    assert f.request.content == 'test'
